requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.1.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
xgboost==2.1.3
joblib==1.4.2
requests==2.32.3
gunicorn==23.0.0
